+++ f3-sidecar/main.go
@@ -22,17 +22,21 @@ func main() {
 	flag.StringVar(&rpcEndpoint, "rpc", "http://127.0.0.1:2345/rpc/v1", "forest RPC endpoint")
 	var f3RpcEndpoint string
 	flag.StringVar(&f3RpcEndpoint, "f3-rpc", "127.0.0.1:23456", "The RPC endpoint F3 sidecar listens on")
+	var initialPowerTable string
+	flag.StringVar(&initialPowerTable, "initial-power-table", "", "The CID of the initial power table")
+	var bootstrapEpoch int64
+	flag.Int64Var(&bootstrapEpoch, "bootstrap", -1, "F3 bootstrap epoch")
 	var finality int64
 	flag.Int64Var(&finality, "finality", 900, "chain finality epochs")
-	var db string
-	flag.StringVar(&db, "db", "f3-db", "path to the f3 database")
+	var root string
+	flag.StringVar(&root, "root", "f3-data", "path to the f3 data directory")
 	var manifestServer string
 	flag.StringVar(&manifestServer, "manifest-server", "", "the peer id of the dynamic manifest server")
 	flag.Parse()
 
 	ctx := context.Background()
 
-	err := run(ctx, rpcEndpoint, f3RpcEndpoint, finality, db, manifestServer)
+	err := run(ctx, rpcEndpoint, f3RpcEndpoint, initialPowerTable, bootstrapEpoch, finality, root, manifestServer)
 	if err != nil {
 		panic(err)
 	}