+++ src/networks/mainnet/mod.rs
@@ -73,6 +73,8 @@ pub static HEIGHT_INFOS: Lazy<HashMap<Height, HeightInfo>> = Lazy::new(|| {
         make_height!(Phoenix, 3_855_480),
         // Tue  6 Aug 12:00:00 UTC 2024
         make_height!(Waffle, 4_154_640, get_bundle_cid("v14.0.0")),
+        // TODO(forest): https://github.com/ChainSafe/forest/issues/4801
+        make_height!(TukTuk, i64::MAX),
     ])
 });