+++ src/rpc/mod.rs
@@ -86,6 +86,7 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::eth::EthGetMessageCidByTransactionHash);
         $callback!(crate::rpc::eth::EthGetStorageAt);
         $callback!(crate::rpc::eth::EthGetTransactionByHash);
+        $callback!(crate::rpc::eth::EthGetTransactionByHashLimited);
         $callback!(crate::rpc::eth::EthGetTransactionCount);
         $callback!(crate::rpc::eth::EthGetTransactionHashByCid);
         $callback!(crate::rpc::eth::EthGetTransactionByBlockNumberAndIndex);
@@ -238,6 +239,8 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::f3::F3GetCertificate);
         $callback!(crate::rpc::f3::F3GetECPowerTable);
         $callback!(crate::rpc::f3::F3GetF3PowerTable);
+        $callback!(crate::rpc::f3::F3IsRunning);
+        $callback!(crate::rpc::f3::F3GetProgress);
         $callback!(crate::rpc::f3::F3GetLatestCertificate);
         $callback!(crate::rpc::f3::F3Participate);
         $callback!(crate::rpc::f3::GetHead);